companyapi/requirements.txt
Django==3.1
django-filter==2.4.0
psycopg2==2.9.5
djangorestframework==3.12
drf-yasg==1.20.0
Pillow==9.4.0
django-environ